+++ geom/line_walker.h
@@ -13,7 +13,7 @@ class LineWalker{
   dLine   line; // line copy
 
   std::vector<double> ls; // lengths from the beginning to every node
-  int     current_n;      // current node
+  size_t  current_n;      // current node
   double  current_l;      // current distance from the current_n
 
 public:
